+++ ui/src/components/attributes-detail/index.ts
@@ -1,0 +1,1 @@
+export * from './attributes-detail';
