+++ ui/src/components/index.ts
@@ -2,4 +2,8 @@ export * from './loading';
 export * from './home-button';
 export * from './image-preview';
 export * from './tile-info';
+export * from './card';
+export * from './accordion-item';
+export * from './input-field-form';
+export * from './attributes-detail';
