+++ ui/src/views/index.ts
@@ -1,3 +1,5 @@
 export * from './home';
 export * from './mint-site';
+export * from './detail';
+export * from './error-screen';
